bktis:mattiar: instead of persist call findbyemail directly in update
the solution to call the findbyemail in persist will never work as the
user to be updated was fetched from repo and stored in cache. the updates
would then be set on the user to be updated in the cache. when searching
for the email address this was always found as it was already on the user
in cache